Moved generate button to separate component
