+++ src/App.jsx
@@ -1,13 +1,13 @@
 import { useEffect, useState } from "react"
 import { Button, ConfigProvider, theme } from "antd"
 import Navbar from "./components/Navbar"
-import Options from "./components/Options"
+import OptionsWrapper from "./components/options/OptionsWrapper"
 import { KeyOutlined } from "@ant-design/icons"
 import { useAnimate, stagger } from "framer-motion"
 import { useSelector, useDispatch } from "react-redux"
-import generatePassword from "./utils/generatePassword"
 import PasswordsList from "./components/PasswordsList"
-import QuickActions from "./components/QuickActions"
+import QuickActions from "./components/actions/QuickActions"
+import GenerateButton from "./components/actions/GenerateButton"
 
 const App = () => {
   const [scope, animate] = useAnimate();
@@ -32,34 +32,7 @@ const App = () => {
     }
   })
 
-  const dispatch = useDispatch();
   const passwords = useSelector(state => state.passwords)
-  const { quantity, length } = useSelector(state => state.options.general);
-  const { uppercase, lowercase, numbers, symbols, separators } = useSelector(state => state.options.include);
-
-  const addPasswords = () => {
-    const passwords = generatePassword({
-      quantity, length, uppercase, lowercase, numbers, symbols, separators
-    })
-
-    dispatch({
-      type: 'passwords/add',
-      payload: passwords,
-    })
-  }
-
-  const [disableGenerateBtn, setDisableGenerateBtn] = useState(false);
-  useEffect(() => {
-    if (
-      (quantity < 1 || quantity > 100) ||
-      (length < 4 || length > 150) ||
-      (!uppercase && !lowercase && !numbers && !symbols)
-    ) {
-      setDisableGenerateBtn(true)
-    } else {
-      setDisableGenerateBtn(false)
-    }
-  }, [quantity, length, uppercase, lowercase, numbers, symbols])
 
   return (
     <ConfigProvider
@@ -76,14 +49,9 @@ const App = () => {
         "
         ref={scope}
       >
-        <Options />
-        <Button
-          type="primary"
-          icon={<KeyOutlined className="text-lg"/>}
-          className="flex items-center bg-primary"
-          onClick={addPasswords}
-          disabled={disableGenerateBtn}
-        >Generate Passwords</Button>
+        <OptionsWrapper />
+        <GenerateButton
+        />
         <QuickActions />
         <PasswordsList
           passwords={passwords}
