+++ src/components/options/GeneralOptions.jsx
